move loader to seperate component
