+++ frontend/src/components/Block.tsx
@@ -1,6 +1,6 @@
 import { Rect, Text, Group, Circle, Arc } from "react-konva";
 import { BlockI } from "../models/block.model";
-import { useEffect, useRef, useState } from "react";
+import { useState } from "react";
 import {
   handleBlockSingleClick,
   handleDragBlockEnd,
@@ -15,6 +15,7 @@ import Link from "./Link";
 import LinkCircleI from "../models/linkcircle.model";
 import { BLOCK_HEIGHT, BLOCK_WIDTH } from "../constants/ui";
 import Konva from "konva";
+import Loader from "./Loader";
 
 interface BlockProps {
   block: BlockI;
@@ -52,24 +53,6 @@ const Block = ({
     setdynamicPosition({ x: e.target.x(), y: e.target.y() });
   };
 
-  const loaderRef = useRef<Konva.Arc>(null);
-
-  useEffect(() => {
-    const angularSpeed = 360;
-    const anim = new Konva.Animation((frame) => {
-      if (frame) {
-        const angleDiff = (frame.timeDiff * angularSpeed) / 1000;
-        loaderRef.current?.rotate(angleDiff);
-      }
-    }, loaderRef.current?.getLayer() || null);
-
-    anim.start();
-
-    return () => {
-      anim.stop();
-    };
-  }, []);
-
   return (
     <Group>
       <Rect
@@ -100,17 +83,11 @@ const Block = ({
         onClick={() => handleBlockSingleClick(block, setBlocks)}
         onDblClick={() => handleDoubleClick(block)}
       />
-      {block.isQueued && (
-        <Arc
-          ref={loaderRef}
-          x={dynamicPosition.x + (BLOCK_WIDTH * 9) / 10}
-          y={dynamicPosition.y + (BLOCK_HEIGHT * 1.5) / 10}
-          fill="darkSlateGray"
-          angle={100}
-          innerRadius={BLOCK_HEIGHT / 20}
-          outerRadius={BLOCK_HEIGHT / 10}
-        />
-      )}
+      <Loader
+        show={block.isQueued}
+        x={dynamicPosition.x + (BLOCK_WIDTH * 9) / 10}
+        y={dynamicPosition.y + (BLOCK_HEIGHT * 1.5) / 10}
+      />
       <Text
         x={dynamicPosition.x}
         y={dynamicPosition.y}
